adding styles for challenge 2 scenarios and outcomes
