+++ src/BadEndPage.js
@@ -11,19 +11,21 @@ function BadEndPage({setDisplay, username, sword, userStats}) {
     }
 
     return(
-        <div>
-            <h1>Well, <span className="username">{username}</span></h1>
-            <p>Looks like this is the end...You made some pretty interesting decisions along the way, but don't worry! If you didn't like the way things panned out I encourage you to try again! Afterall, adventurers never give up!</p>
-            <p>Don't forget to take a look at how you did!</p>
+        <div className='displayContainer BadEndPageContainer'>
+            <div className='contentContainer'>
+                <h1>Well, <span className="username">{username}</span></h1>
+                <p>Looks like this is the end...You made some pretty interesting decisions along the way, but don't worry! If you didn't like the way things panned out I encourage you to try again! Afterall, adventurers never give up!</p>
+                <p>Don't forget to take a look at how you did!</p>
 
-            <button
-            onClick={() =>{
-                    setDisplay(<UserStatsPage username={username} setDisplay={setDisplay} sword={sword} userStats={userStats}/>)
-                    userStats.push(`You died a miserable and/or painful death`);
-                    addUser(userStats);
+                <button
+                onClick={() =>{
+                        setDisplay(<UserStatsPage username={username} setDisplay={setDisplay} sword={sword} userStats={userStats}/>)
+                        userStats.push(`You died a miserable and/or painful death`);
+                        addUser(userStats);
+                        }
                     }
-                }
-            >User Stats</button>
+                >User Stats</button>
+            </div>
         </div>
     )
 }
