+++ src/finalChallenge/ChallengeScenario1/Scenario1Outcome2/Scenario1Outcome2.js
@@ -2,17 +2,19 @@ import BadEndPage from "../../../BadEndPage.js";
 
 function Scenario1Outcome2({setDisplay, username, sword, userStats}) {
     return(
-        <div>
-            <h1>Objective: Say your prayers..</h1>
-            <p>Your vision fades to darkness as the dragons mouth grows nearer. Your final thoughts are will I taste good? I hope not..</p>
+        <div className='displayContainer Scenario1Outcome2'>
+            <div className='contentContainer'>
+                <h1>Objective: Say your prayers..</h1>
+                <p>Your vision fades to darkness as the dragons mouth grows nearer. Your final thoughts are will I taste good? I hope not..</p>
 
-            <button
-            onClick={() =>{
-                    setDisplay(<BadEndPage username={username} setDisplay={setDisplay} sword={sword} userStats={userStats}/>)
-                    userStats.push(`You were eaten whole!`)
+                <button
+                onClick={() =>{
+                        setDisplay(<BadEndPage username={username} setDisplay={setDisplay} sword={sword} userStats={userStats}/>)
+                        userStats.push(`You were eaten whole!`)
+                        }
                     }
-                }
-            >Continue</button>
+                >Continue</button>
+            </div>
         </div>
     )
 }
