+++ src/taskSword/swordOptions/SwordOption3.js
@@ -2,34 +2,37 @@ import MainPage from "../../MainPage.js";
 import woodenSword from "../../assets/swords/wooden_sword.png";
 
 function SwordOption3({setDisplay, username, sword, userStats}) {
-    // let missingLetter;
-    // function nameScrambled(username) {
-    //     let newName = [...username];
-    //     const index = Math.floor(Math.random() * newName.length);
-    //     missingLetter = newName[index];
-    //     newName.splice(index, 1);
-    //     newName[0] = newName[0].toUpperCase();
-    //     newName.join('');
-    //     return newName;
-    // }
-    // setSword('Wooden Sword');
+    let missingLetter;
+    let newName;
+    function nameScrambled(username) {
+        let alteredName = [...username];
+        const index = Math.floor(Math.random() * alteredName.length);
+        missingLetter = alteredName[index];
+        alteredName.splice(index, 1);
+        alteredName.join('');
+        return alteredName;
+    }
+    newName = nameScrambled(username);
+
     return(
-        <div>
-            <h1>Objective: Obtain the Dragon Long Sword</h1>
+        <div className='displayContainer SwordOption1Container'>
+            <div className='contentContainer'>
+                <h1>Objective: Obtain A Weapon</h1>
 
-            {/* Smiley is absolutely wrong. This is unbelievable..how would you feel just called you '{nameScrambled}' and took out the '{missingLetter}' from your name ? */}
-            <p> Smiley is absolutely wrong. This is unbelievable..and you call yourself an adventurer? What's your rank anyhow? C? D? Oh oh, maybe F! You lot like to run around playing with weapons that you have no business carrying around. Here..take this and when you acquire some respect for our work, then maybe I'll make you a <span className="dragonLongSword">Dragon Long Sword</span>. For now, walk around with this <span className="woodenSword">Wooden Sword</span>.</p>
+                <p>Smiley is absolutely wrong. This is unbelievable..how would you feel if I took out the '{missingLetter}' from your name and just called you '{newName}'?</p>
+                <p>..and you call yourself an adventurer? What's your rank anyhow? C? D? Oh oh, maybe F! You lot like to run around playing with weapons that you have no business carrying around. Here..take this and when you acquire some respect for our work, then maybe I'll make you a <span className="dragonLongSword">Dragon Long Sword</span>. For now, walk around with this <span className="woodenSword">Wooden Sword</span>.</p>
 
-            <div className="imageContainer">
-                <img src={woodenSword} alt="wooden sword"/>
-            </div>
-            
-            <button
-                onClick={() => {
-                    setDisplay(<MainPage username={username} setDisplay={setDisplay} sword={sword} userStats={userStats} />)
+                <div className="imageContainer woodenSwordContainer">
+                    <img src={woodenSword} alt="wooden sword"/>
+                </div>
+                
+                <button
+                    onClick={() => {
+                        setDisplay(<MainPage username={username} setDisplay={setDisplay} sword={sword} userStats={userStats} />)
+                        }
                     }
-                }
-            >Return to the Main page</button>
+                >Return to the Main page</button>
+            </div>
         </div>
     )
 }
